Adapt shared as discussed
